Basically working pattern for interpreting dictation results
Currently just uses a flat-file, will eventually need to allow the
user to define/edit these patterns. The patterns for reduction of
white-space are currently crude/ugly. A mechanism for handling
more complex patterns, such as "number ..." where you want to hook
in a specific mode/sub-interpretation would be nice.